C. subellipsoidea, D. salina, M. pusilla and O. lucimarinus' packages
Packages have been generated but not tested, I have a problem when I try to install them because in windows there is a different way to install local packages (or at least that what I've read). So I'm still trying to figure out the way to install them. TO DO: ask Mercedes about the other algae that we can include.